Support Embed on EventType Edit, Team Link Fix and Mobile unsupported
